% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KFinReport-package.R
\docType{package}
\name{KFinReport-package}
\alias{KFinReport}
\alias{KFinReport-package}
\title{KFinReport: Handling Korea's Corporate Filings from DART}
\description{
This package utilizes Korea's Repository of Corporate Filings (DART).
    It's goal is to assist users to browse through the repository via R. 
    Users are able to retrieve corporate filings as data.
    Upon retrieval, users can select the corporate of interest, the time span,
    types of report, etc.
}
\author{
\strong{Maintainer}: Sang Sun Ahn \email{sangsun.ahn@gmail.com}

}
\keyword{internal}
