% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EarningReports.R
\name{report}
\alias{report}
\title{Get a Single Companies' All Financial Statement}
\usage{
report(crtfc_key, corp_code, bsns_year, reprt_code, fs_div)
}
\arguments{
\item{crtfc_key}{is API certification key issued by openDART.}

\item{corp_code}{is corporation code of the company you want to look for.}

\item{bsns_year}{is business year to look for. Default is current year.}

\item{reprt_code}{is report code to look for.}

\item{fs_div}{is whether financial statements are consolidated or not. Default is OFS.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Get Financial Statement for korea finance market
}
